+++ src/types/env.ts
@@ -1,0 +1,5 @@
+export type Environment = {
+  Bindings: {
+    CLOUDINARY_CLOUD_NAME: string
+  }
+}