+++ src/types/graphql.ts
@@ -1,0 +1,10 @@
+export interface GraphQLEnsDomainResponse {
+  data: {
+    domains: Array<Domain>
+  }
+}
+
+interface Domain {
+  id: string
+  name: string
+}